add new riff
